Create a Util class for testing
