+++ test/Util.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Test;
+
+/**
+ * Some usefull methods for testing
+ *
+ *
+ * @author Johann SERVOIRE @Johann-S
+ * @since 0.1
+ */
+class Util
+{
+    /**
+     * Call protected/private method of a class.
+     *
+     * @param object &$object    Instantiated object that we will run method on.
+     * @param string $methodName Method name to call
+     * @param array  $parameters Array of parameters to pass into method.
+     *
+     * @return mixed Method return.
+     */
+    public function invokeMethod(&$object, $methodName, array $parameters = array())
+    {
+        $reflection = new \ReflectionClass(get_class($object));
+        $method = $reflection->getMethod($methodName);
+        $method->setAccessible(true);
+        return $method->invokeArgs($object, $parameters);
+    }
+}
